+++ pyemma/msm/estimators/augmented_msm.py
@@ -39,7 +39,8 @@ class AugmentedMarkovModel(MaximumLikelihoodMSM):
     def __init__(self, lag=1, count_mode='sliding', connectivity='largest',
                  dt_traj='1 step',
                  E=None, m=None, w=None, eps=0.05, support_ci=1.00, maxiter=500, max_cache=3000,
-                 mincount_connectivity='1/n'):
+                 mincount_connectivity='1/n', core_set=None, milestoning_method='last_core'):
+
         r"""Maximum likelihood estimator for AMMs given discrete trajectory statistics and expectation values from experiments
 
         Parameters
@@ -139,7 +140,8 @@ class AugmentedMarkovModel(MaximumLikelihoodMSM):
         super(AugmentedMarkovModel, self).__init__(lag=lag, reversible=True, count_mode=count_mode, sparse=False,
                                                    connectivity=connectivity, dt_traj=dt_traj, score_method=None,
                                                    score_k=None, mincount_connectivity=mincount_connectivity,
-                                                   maxiter=maxiter)
+                                                   maxiter=maxiter, core_set=core_set,
+                                                   milestoning_method=milestoning_method)
 
         self.E = E
         if E is not None:
@@ -329,9 +331,6 @@ class AugmentedMarkovModel(MaximumLikelihoodMSM):
         if _np.size(self.active_set) == 0:
             raise RuntimeError('Active set is empty. Cannot estimate AMM.')
 
-        from pyemma.util.discrete_trajectories import index_states
-        self._active_state_indexes = index_states(dtrajs, subset=self.active_set)
-
         # active count matrix and number of states
         self._C_active = dtrajstats.count_matrix(subset=self.active_set)
         self._nstates = self._C_active.shape[0]
@@ -342,7 +341,8 @@ class AugmentedMarkovModel(MaximumLikelihoodMSM):
         self._full2active[self.active_set] = _np.arange(len(self.active_set))
 
         # slice out active states from E matrix
-        _dset = list(set(_np.concatenate(dtrajs)))
+
+        _dset = list(set(_np.concatenate(self._dtrajs_full)))
         _rras = [_dset.index(s) for s in self.active_set]
         self.E_active = self.E[_rras]
 
@@ -474,7 +474,6 @@ class AugmentedMarkovModel(MaximumLikelihoodMSM):
 
         _P = msmest.tmatrix(self._C_active, reversible=True, mu=self._pihat)
 
-        self._dtrajs_full = dtrajs
         self._connected_sets = msmest.connected_sets(self._C_full)
         self.set_model_params(P=_P, pi=self._pihat, reversible=True,
                               dt_model=self.timestep_traj.get_scaled(self.lag))