+++ pyemma/msm/estimators/maximum_likelihood_msm.py
@@ -39,7 +39,7 @@ class MaximumLikelihoodMSM(_MSMEstimator):
                  count_mode='sliding', sparse=False,
                  connectivity='largest', dt_traj='1 step', maxiter=1000000,
                  maxerr=1e-8, score_method='VAMP2', score_k=10,
-                 mincount_connectivity='1/n'):
+                 mincount_connectivity='1/n', core_set=None, milestoning_method='last_core'):
         r"""Maximum likelihood estimator for MSMs given discrete trajectory statistics
 
         Parameters
@@ -147,6 +147,18 @@ class MaximumLikelihoodMSM(_MSMEstimator):
             may thus separate the resulting transition matrix. The default
             evaluates to 1/nstates.
 
+        core_set : None (default) or array like, dtype=int
+            Definition of core set for milestoning MSMs.
+            If set to None, replaces state -1 (if found in discrete trajectories) and
+            performs milestone counting. No effect for Voronoi-discretized trajectories (default).
+            If a list or np.ndarray is supplied, discrete trajectories will be assigned
+            accordingly.
+
+        milestoning_method : str
+            Method to use for counting transitions in trajectories with unassigned frames.
+            Currently available:
+            |  'last_core',   assigns unassigned frames to last visited core
+
         References
         ----------
         .. [1] H. Wu and F. Noe: Variational approach for learning Markov processes from time series data
@@ -156,7 +168,8 @@ class MaximumLikelihoodMSM(_MSMEstimator):
         super(MaximumLikelihoodMSM, self).__init__(lag=lag, reversible=reversible, count_mode=count_mode,
                                                    sparse=sparse, connectivity=connectivity, dt_traj=dt_traj,
                                                    score_method=score_method, score_k=score_k,
-                                                   mincount_connectivity=mincount_connectivity)
+                                                   mincount_connectivity=mincount_connectivity,
+                                                   core_set=core_set, milestoning_method=milestoning_method)
 
         self.statdist_constraint = _types.ensure_ndarray_or_None(statdist_constraint, ndim=None, kind='numeric')
         if self.statdist_constraint is not None:  # renormalize
@@ -196,6 +209,20 @@ class MaximumLikelihoodMSM(_MSMEstimator):
         dtrajstats = self._get_dtraj_stats(dtrajs)
         self._C_full = dtrajstats.count_matrix()  # full count matrix
         self._nstates_full = self._C_full.shape[0]  # number of states
+
+        # check for consistency between statdist constraints and core set
+        if self.core_set is not None and self.statdist_constraint is not None:
+            if len(self.core_set) != len(self.statdist_constraint):
+                raise ValueError('Number of core sets and stationary distribution '
+                                 'constraints do not match.')
+
+            # rewrite statdist constraints to full set for compatibility reasons
+            #TODO: find a more consistent way of dealing with this
+            import copy
+            _stdist_constr_coreset = copy.deepcopy(self.statdist_constraint)
+            self.statdist_constraint = _np.zeros(self._nstates_full)
+            self.statdist_constraint[self.core_set] = _stdist_constr_coreset
+
 
         # set active set. This is at the same time a mapping from active to full
         if self.connectivity == 'largest':
@@ -275,7 +302,6 @@ class MaximumLikelihoodMSM(_MSMEstimator):
 
         # Done. We set our own model parameters, so this estimator is
         # equal to the estimated model.
-        self._dtrajs_full = dtrajs
         self._connected_sets = dtrajstats.connected_sets
         self.set_model_params(P=P, pi=statdist_active, reversible=self.reversible,
                               dt_model=self.timestep_traj.get_scaled(self.lag))